Make generic TableDropdown component and use it on occurrence table actions dropdown
