+++ src/common/components/tableDropdown/tableDropdown.module.scss
@@ -1,0 +1,94 @@
+.tableDropdown {
+  --background-color-menu: var(--color-white);
+  --background-color-menu-item: var(--color-white);
+  --background-color-menu-item-focus: var(--color-black-10);
+
+  --border-color-button: var(--color-white);
+  --border-color-button-focus: var(--color-black);
+
+  --color-button: var(--color-black);
+
+  --fontsize-button: var(--fontsize-body-medium);
+  --fontsize-menu-item: var(--fontsize-body-medium);
+
+  --spacing-button: var(--spacing-xs);
+  --spacing-menu-item: var(--spacing-xs);
+
+  --z-index-menu: 1000;
+}
+.tableDropdown {
+  position: relative;
+  align-self: center;
+  margin: 0 calc(0px - var(--spacing-button));
+
+  .toggleButton {
+    display: flex;
+    font-size: var(--fontsize-button);
+    color: var(--color-button);
+    font-weight: normal;
+    text-align: left;
+    align-items: center;
+    padding: var(--spacing-button);
+    border: 2px solid var(--border-color-button);
+    width: 100%;
+
+    span {
+      flex: 1 1 0%;
+    }
+
+    .iconAngleDown {
+      margin-left: var(--spacing-xs);
+    }
+
+    &:focus {
+      border: 2px solid var(--border-color-button-focus);
+    }
+  }
+
+  .dropdownMenu {
+    display: none;
+  }
+
+  &.isMenuOpen {
+    .toggleButton {
+      .iconAngleDown {
+        transform: rotate(180deg);
+      }
+    }
+
+    .dropdownMenu {
+      display: block;
+      position: absolute;
+      margin: 0;
+      padding: 0;
+      list-style: none;
+      top: 100%;
+      right: 0;
+      background-color: var(--background-color-menu);
+      box-shadow: 0px 0px 6px 0px rgba(0, 0, 0, 0.2);
+      z-index: var(--z-index-menu);
+
+      li {
+        background-color: var(--background-color-menu-item);
+        padding: var(--spacing-menu-item);
+        font-size: var(--fontsize-menu-item);
+        cursor: pointer;
+        white-space: nowrap;
+        display: flex;
+        align-items: center;
+
+        &.isFocused {
+          background-color: var(--background-color-menu-item-focus);
+        }
+
+        &:hover {
+          text-decoration: underline;
+        }
+
+        svg {
+          margin-right: var(--spacing-xs);
+        }
+      }
+    }
+  }
+}
